test: update tests and goldens
